create email_routes.py for code refactor
