+++ backend/db/utils/user_utils.py
@@ -6,7 +6,7 @@ from datetime import datetime, timedelta, timezone
 logger = logging.getLogger(__name__)
 
 def user_exists(user) -> bool:
-    from main import engine
+    from database import engine
     """
     Checks if user is already in the database
 
@@ -23,7 +23,7 @@ def add_user(user) -> Users:
     Writes user data to the users model
 
     """
-    from main import engine
+    from database import engine
     with Session(engine) as session:
         # Check if the user already exists in the database
         existing_user = session.exec(select(Users).where(Users.user_id == user.user_id)).first()
